convertingonthefly:

+++ example_kvcache_organization.py
@@ -453,7 +453,7 @@ def Vanilla_specu_dectesting3(tokenizer,
         
         print("target_model_logits dtype: {}".format(target_model_logits.dtype)) 
         print("outputs.logits dtype: {}".format(outputs.logits.dtype)) 
-        torch.allclose(target_model_logits, outputs.logits[:, -1, :]) # check if the two logits are the same 
+        torch.allclose(target_model_logits, outputs.logits[:, -1, :].to(torch.bfloat16))  # check if the two logits are the same 
 
     count = 0
     verify_probs = []
